updates to comments
