+++ app/components/header/hooks/use-scroll-direction.tsx
@@ -1,0 +1,25 @@
+import { useEffect, useState } from 'react';
+
+export function useScrollDirection(): boolean {
+  const [scrollingUp, setScrollingUp] = useState(false);
+  const [scrollPosition, setScrollPosition] = useState(0);
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const currentScrollPosition = window.scrollY;
+      if (currentScrollPosition < scrollPosition) {
+        setScrollingUp(true);
+      } else if (currentScrollPosition > scrollPosition) {
+        setScrollingUp(false);
+      }
+      setScrollPosition(currentScrollPosition)
+    }
+
+    window.addEventListener('scroll', handleScroll);
+
+    return () => {
+      window.removeEventListener('scroll', handleScroll)
+    };
+  }, [scrollPosition])
+  return scrollingUp;
+}
